captureの停止ボタン実装
