+++ src/App.js
@@ -3,91 +3,8 @@ import "./styles.css";
 import "@tensorflow/tfjs";
 import * as handPoseDetection from "@tensorflow-models/hand-pose-detection";
 import Webcam from "react-webcam";
-import { Canvas, useFrame } from "@react-three/fiber";
-
-const scale = (point) => -(point * 800) / 50;
-
-const Finger = ({ predictionsRef, fingerIndex }) => {
-  const jointBottom = React.useRef();
-  const jointMiddleBottom = React.useRef();
-  const jointMiddleTop = React.useRef();
-  const jointTop = React.useRef();
-
-  const updateJoint = (point, thumb) => {
-    thumb.current.position.x = -scale(point.x);
-    thumb.current.position.y = scale(point.y);
-    thumb.current.position.z = scale(point.z);
-  };
-
-  useFrame(() => {
-    if (predictionsRef.current.length) {
-      updateJoint(
-        predictionsRef.current[0].keypoints3D[Number(fingerIndex)],
-        jointBottom
-      );
-      updateJoint(
-        predictionsRef.current[0].keypoints3D[Number(fingerIndex) + 1],
-        jointMiddleBottom
-      );
-      updateJoint(
-        predictionsRef.current[0].keypoints3D[Number(fingerIndex) + 2],
-        jointMiddleTop
-      );
-      updateJoint(
-        predictionsRef.current[0].keypoints3D[Number(fingerIndex) + 3],
-        jointTop
-      );
-    }
-  });
-
-  return (
-    <>
-      <mesh castShadow receiveShadow ref={jointBottom} scale={[1, 1, 1]}>
-        <sphereBufferGeometry attach="geometry" args={[0.1, 32, 32]} />
-        <meshStandardMaterial attach="material" color="white" />
-      </mesh>
-      <mesh castShadow receiveShadow ref={jointMiddleBottom} scale={[1, 1, 1]}>
-        <sphereBufferGeometry attach="geometry" args={[0.1, 32, 32]} />
-        <meshStandardMaterial attach="material" color="white" />
-      </mesh>
-      <mesh castShadow receiveShadow ref={jointMiddleTop} scale={[1, 1, 1]}>
-        <sphereBufferGeometry attach="geometry" args={[0.1, 32, 32]} />
-        <meshStandardMaterial attach="material" color="white" />
-      </mesh>
-      <mesh castShadow receiveShadow ref={jointTop} scale={[1, 1, 1]}>
-        <sphereBufferGeometry attach="geometry" args={[0.1, 32, 32]} />
-        <meshStandardMaterial attach="material" color="#eb3b5a" />
-      </mesh>
-    </>
-  );
-};
-
-const Hand = ({ predictionsRef }) => {
-  const palm = React.useRef();
-
-  useFrame(() => {
-    if (predictionsRef.current.length) {
-      const point = predictionsRef.current[0].keypoints3D[0];
-      palm.current.position.x = -scale(point.x);
-      palm.current.position.y = scale(point.y);
-      palm.current.position.z = scale(point.z);
-    }
-  });
-
-  return (
-    <>
-      <mesh castShadow receiveShadow ref={palm} scale={[1, 1, 1]}>
-        <sphereBufferGeometry attach="geometry" args={[0.1, 32, 32]} />
-        <meshStandardMaterial attach="material" color="#3867d6" />
-      </mesh>
-      <Finger predictionsRef={predictionsRef} fingerIndex="1" />
-      <Finger predictionsRef={predictionsRef} fingerIndex="5" />
-      <Finger predictionsRef={predictionsRef} fingerIndex="9" />
-      <Finger predictionsRef={predictionsRef} fingerIndex="13" />
-      <Finger predictionsRef={predictionsRef} fingerIndex="17" />
-    </>
-  );
-};
+import { Canvas } from "@react-three/fiber";
+import { Hand } from "./Hand";
 
 export default function App() {
   const webcamRef = React.useRef(null);
@@ -105,14 +22,14 @@ export default function App() {
       if (predictions) {
         predictionsRef.current = predictions;
       }
-
-      if (!ready) {
-        setReady(true);
-      }
     }
 
-    requestRef.current = requestAnimationFrame(capture);
-  }, [webcamRef, ready]);
+    if (ready) {
+      requestRef.current = requestAnimationFrame(capture);
+    } else {
+      requestRef.current = null;
+    }
+  }, [ready]);
 
   React.useEffect(() => {
     const load = async () => {
@@ -129,6 +46,17 @@ export default function App() {
 
     load();
   }, []);
+
+  React.useEffect(() => {
+    if (ready) {
+      requestRef.current = requestAnimationFrame(capture);
+    } else {
+      if (requestRef.current) {
+        console.log("cancel");
+        cancelAnimationFrame(requestRef.current);
+      }
+    }
+  }, [ready]);
 
   return (
     <>
@@ -166,34 +94,28 @@ export default function App() {
           screenshotFormat="image/jpeg"
         />
       </div>
-      {!ready && (
-        <div
-          style={{
-            backgroundColor: "rgba(23,32,23,0.3)",
-            position: "absolute",
-            color: "white",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            cursor: "pointer",
-            right: 0,
-            top: 0,
-            left: 0,
-            bottom: 0,
+      <div
+        style={{
+          backgroundColor: "rgba(23,32,23,0.3)",
+          position: "absolute",
+          color: "white",
+          display: "flex",
+          cursor: "pointer",
+          top: 10,
+          left: 10,
+        }}
+      >
+        <button
+          onClick={() => {
+            setReady(!ready);
           }}
         >
-          <button
-            onClick={() => {
-              requestRef.current = requestAnimationFrame(capture);
-            }}
-          >
-            Start hand tracking{" "}
-            <span role="img" aria-label="Start">
-              🖐
-            </span>
-          </button>
-        </div>
-      )}
+          Start hand tracking{" "}
+          <span role="img" aria-label="Start">
+            🖐
+          </span>
+        </button>
+      </div>
     </>
   );
 }
